refactor(atls-ui-generators): extract lib from scripts
