+++ ui-admin/icons/lib/check-generated/errors/command-failed.error.ts
