+++ ui-admin/icons/lib/check-generated/filesystem/collect-files.ts
