+++ ui-admin/icons/lib/check-generated/filesystem/is-directory.ts
