+++ ui-admin/icons/lib/check-generated/filesystem/is-file.ts
