+++ ui-admin/icons/lib/check-generated/generation/generate-replacements.ts
@@ -5,7 +5,7 @@ import { iconsPathArgument }           from '../constants.js'
 import { workspaceRootPath }           from '../constants.js'
 import { workspaceArgument }           from '../constants.js'
 import { yarnCommand }                 from '../constants.js'
-import { runCommand }                  from './run-command.js'
+import { runCommand }                  from '../process/run-command.js'
 
 export const generateReplacements = async (svgPath: string, outputPath: string): Promise<void> => {
   await runCommand(
