+++ ui-admin/icons/lib/check-generated/interfaces.ts
